Update paths, histcontrol, and WSL support
Changed JAVA_HOME to point to the default path created by Ubuntu
default-jdk-headless package.

Added additonal paths in the home directory to the PATH env variable.

Added a check to see if the Windows Subsystem for Linux (WSL) is
activated, and if so add an entry to the PATH env variable and ensure
cron is running.

Set histcontrol to ignore whitespace so sensitive input on the command
line are not saved in history.